# application/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "API for managing a mechanic shop, including customers, employees, service tickets, inventory, and services"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions: 
  bearerAuth: # name of the security scheme
    type: apiKey
    name: Authorization
    in: header
    description: "JWT token, prefixed with 'Bearer '"
parameters:
  PageParam:
    name: page
    in: query
    type: integer
    required: false
    default: 1
  LimitParam:
    name: limit
    in: query
    type: integer
    required: false
    default: 10
  SortByParam:
    name: sort_by
    in: query
    type: string
    required: false
  SortOrderParam:
    name: sort_order
    in: query
    type: string
    enum: [asc, desc]
    required: false
  StatusParam:
    name: status
    in: query
    type: string
    enum: [open, in_progress, closed]
    required: false
  NameParam:
    name: name
    in: query
    type: string
    required: false
  EmailParam:
    name: email
    in: query
    type: string
    required: false
...
  SearchParam:
    name: search
    in: query
    type: string
    required: false
paths:
...
  /customers/login:
    post:
      tags:
        - Customer
      summary: "Customer login (generates token)"
      description: "Logs in a customer using email and password. generates and return a customer specific token. Rate limited to 3 requests per 10 minutes."
      parameters:
        - in: body
          name: body
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginSuccessResponse"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        429:
          description: "Too Many Requests"
          produces:
            - text/html
          schema:
            type: string
            example: "<!doctype html>\n<html lang=en>\n<title>429 Too Many Requests</title>\n<h1>Too Many Requests</h1>\n<p>3 per 10 minute</p>"
          examples:
            text/html: |
              <!doctype html>
              <html lang=en>
              <title>429 Too Many Requests</title>
              <h1>Too Many Requests</h1>
              <p>3 per 10 minute</p>
  /customers:
    post:
      tags:
        - Customer
      summary: "Create new customer"
      description: "Registers a new customer. Rate limited to 3 requests per hour."
      parameters:
        - in: body
          name: customer
          description: "Information required to create a new Customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        409:
          description: "Email already exists"
          schema:
            $ref: "#/definitions/CustomerConflictResponse"
        429:
          description: "Too Many Requests"
          produces:
            - text/html
          schema:
            type: string
          examples:
            text/html: |
              <!doctype html>
              <html lang=en>
              <title>429 Too Many Requests</title>
              <h1>Too Many Requests</h1>
              <p>3 per 1 hour</p>
        500:
          $ref: "#/definitions/ServerErrorResponse"
  /customers/me:
    get:
      tags:
        - Customer
      summary: "Get current authenticated customer's profile"
      description: "Returns the profile of the authenticated customer using their JWT token. Rate limit: 10 requests per minute per user."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer profile retrieved successfully"
          schema:
            $ref: "#/definitions/CustomerSuccessResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden - invalid role"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        429:
          description: "Too Many Requests"
          schema:
            $ref: "#/definitions/RateLimitResponse"
  /customers/me/tickets:
    get:
      tags:
        - Customer
      summary: "Get current customer's service tickets"
      description: "Returns the authenticated customer's service tickets with pagination and optional status filtering."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - $ref: "#/parameters/StatusParam"
      responses:
        200:
          description: "Tickets retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              data:
                type: array
                items:
                  $ref: "#/definitions/CreateServiceTicketResponse"
              meta:
                $ref: "#/definitions/PaginationMeta"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
  /customers/{customer_id}:
    patch:
      tags:
        - Customer
      summary: "Update customer's basic information"
      description: "Updates name, email, or phone of the authenticated customer. Rate limited to 20 requests per hour. Only the authenticated customer can update their own profile"
      security:
        - bearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: "ID of the customer to update (must match authenticated user)"
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/CustomerSuccessResponse"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Not found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
        422:
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/UnprocessableEntityResponse"
        429:
          description: "Too Many Requests"
          schema:
            $ref: "#/definitions/RateLimitResponse"
  /customers/me/update-password:
    patch:
      tags:
        - Customer
      summary: "Update customer's password"
      description: "Updates the authenticated customer's password after verifying the current password."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/PasswordUpdatePayload"
      responses:
        200:
          description: "Password updated successfully"
          schema:
            $ref: "#/definitions/SuccessResponse"
        400:
          description: "Validation error"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Missing required fields"
            examples:
              missing_fields:
                status: "error"
                message: "Missing required fields"
              invalid_password:
                status: "error"
                message: "Password must be at least 8 characters, including letters and numbers."
        401:
          description: "Current password incorrect"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Current password is incorrect"
        403:
          description: "Forbidden - invalid role"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          $ref: "#/definitions/NotFoundResponse"

#MARK: employees paths
...
  /employees/customers:
    get:
      tags:
        - Employee
      summary: "Get all customers"
      description: "Admin/employee view of all customers with pagination, filtering, sorting, and search."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - $ref: "#/parameters/NameParam"
        - $ref: "#/parameters/EmailParam"
        - $ref: "#/parameters/SearchParam"
      responses:
        200:
          description: "Customers retrieved successfully"
          schema:
            $ref: "#/definitions/GetCustomersResponse"
        400:
          description: "Invalid query parameter"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Invalid query parameter"
          examples:
            application/json:
              status: "error"
              message: "Invalid query parameter"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
  /employees/customers/{customer_id}:
    get:
      tags:
        - Employee
      summary: "Get customer by ID"
      description: "Returns a specific customer by ID. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: "ID of the customer to retrieve"
      responses:
        200:
          description: "Customer retrieved successfully"
          schema:
            allOf:
              - $ref: "#/definitions/BaseSuccessResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/definitions/CustomerProfileData"
          examples:
            application/json:
              status: "success"
              message: "Operation successful"
              data:
                id: 1
                name: "zambak"
                email: "zamzam@gmail.com"
                phone: "2908763453"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Customer not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Customer not found"
          examples:
            application/json:
              status: "error"
              message: "Customer not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    
    patch:
      tags:
        - Employee
      summary: "Update customer as employee"
      description: "Allows an employee to update a customer's information. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
          description: "ID of the customer to update"
        - in: body
          name: body
          required: true
          description: "Fields to update. Leave empty to return existing profile."
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            allOf:
              - $ref: "#/definitions/BaseSuccessResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/definitions/CustomerProfileData"
          examples:
            application/json:
              status: "success"
              message: "Operation successful"
              data:
                email: "lily@gmail.com"
                id: 4
                name: "lilybun"
                phone: "5457897583"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Customer not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Customer not found"
          examples:
            application/json:
              status: "error"
              message: "Customer not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
definitions:
  BaseSuccessResponse:
    type: object
    properties:
      status:
        type: string
        example: "success"
      message:
        type: string
        example: "Operation successful"
      data:
        type: object
        description: "Response payload"

  # Common Response Definitions
  ValidationErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Invalid input"
      errors:
        type: object
        properties:
          field_name:
            type: array
            items:
              type: string
            example: ["Missing data for required field."]
    examples:
      application/json:
        status: "error"
        message: "Invalid input"
  UnauthorizedResponse:
    type: object
    properties:
      message:
        type: string
        example: "Authorization header is required"
    examples:
      application/json:
        message: "Authorization header is required"
  ForbiddenResponse:
    type: object
    properties:
      message:
        type: string
        example: "Forbidden - invalid role"
    examples:
      application/json:
        message: "Forbidden - invalid role"
  NotFoundResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Resource not found"
    examples:
      application/json:
        status: "error"
        message: "Resource not found"
  ServerErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Internal server error"
    examples:
      application/json:
        status: "error"
        message: "Internal server error"
  RateLimitResponse:
    type: string
    example: "<!doctype html>\n<html lang=en>\n<title>429 Too Many Requests</title>\n<h1>Too Many Requests</h1>\n<p>3 per 1 hour</p>"
    examples:
      text/html: |
        <!doctype html>
        <html lang=en>
        <title>429 Too Many Requests</title>
        <h1>Too Many Requests</h1>
        <p>3 per 1 hour</p>
  UnprocessableEntityResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Validation errors occurred"
      errors:
        type: object
        properties:
          field_name:
            type: array
            items:
              type: string
            example: ["Invalid format"]
    examples:
      application/json:
        status: "error"
        message: "Validation errors occurred"
        errors:
          phone: ["Invalid phone format"]
  PaginationMeta:
    type: object
    properties:
      pagination:
        type: object
        properties:
          has_next:
            type: boolean
            example: false
          has_prev:
            type: boolean
            example: false
          limit:
            type: integer
            example: 10
          page:
            type: integer
            example: 1
          total_items:
            type: integer
            example: 1
          total_pages:
            type: integer
            example: 1
  SuccessResponse:
    allOf:
      - $ref: "#/definitions/BaseSuccessResponse"
      - type: object
        examples:
          application/json:
            status: "success"
            message: "Operation successful"

#MARK: Login Related
  LoginCredentials:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
      password:
        type: string
...
  LoginSuccessResponse:
    allOf:
      - $ref: "#/definitions/BaseSuccessResponse"
      - type: object
        properties:
          data:
            type: object
            properties:
              token:
                type: string
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        examples:
          application/json:
            status: "success"
            message: "Login successful"
            data:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

#MARK: customers definitions
  CreateCustomerPayload:
    type: object
    required:
      - name
      - email
      - phone
      - password
    properties:
      name:
        type: string
        maxLength: 50
      email:
        type: string
        format: email
      phone:
        type: string
        pattern: "^\\d{10}$"
        description: "Must be exactly 10 digits"
      password:
        type: string
        format: password
        description: "Must be at least 8 characters long, including letters and numbers"
  CreateCustomerResponse:
    allOf:
      - $ref: "#/definitions/BaseSuccessResponse"
      - type: object
        properties:
          data:
            $ref: "#/definitions/CustomerProfileData"
        examples:
          application/json:
            status: "success"
            message: "Customer created successfully"
            data:
              id: 9
              name: "hugo"
              email: "hugo@gmail.com"
              phone: "9667766551"
  CustomerProfileData:
    type: object
    properties:
      id:
        type: integer
        example: 9
      name:
        type: string
        example: "hugo"
      email:
        type: string
        format: email
        example: "hugo@gmail.com"
      phone:
        type: string
        pattern: "^\\d{10}$"
        example: "9667766551"
  CustomerSuccessResponse:
    allOf:
      - $ref: "#/definitions/BaseSuccessResponse"
      - type: object
        properties:
          data:
            $ref: "#/definitions/CustomerProfileData"
        examples:
          application/json:
            status: "success"
            message: "Operation successful"
            data:
              id: 9
              name: "hugo"
              email: "hugo@gmail.com"
              phone: "9667766551"
  CustomerConflictResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Customer already exists"
      details:
        type: object
        properties:
          customer:
            $ref: "#/definitions/CustomerProfileData"
    examples:
      application/json:
        status: "error"
        message: "Customer already exists"
        details:
          customer:
            id: 9
            name: "hugo"
            email: "hugo@gmail.com"
            phone: "9667766551"
...
  UpdateCustomerPayload:
    type: object
    properties:
      name:
        type: string
        maxLength: 50
      email:
        type: string
        format: email
      phone:
        type: string
        pattern: "^\\d{10}$"
        description: "Must be exactly 10 digits"    
  PasswordUpdatePayload:
    type: object
    required:
      - current_password
      - new_password
    properties:
      current_password:
        type: string
      new_password:
        type: string
        description: "Must be at least 8 characters, including letters and numbers"

#MARK: employee definitions
...
  GetCustomersResponse:
    type: object
    properties:
      status:
        type: string
        example: "success"
      message:
        type: string
        example: "Operation successful"
      meta:
        type: object
        properties:
          pagination:
            $ref: "#/definitions/PaginationMeta"
      data:
        type: array
        items:
          $ref: "#/definitions/CustomerProfileData"
    examples:
      application/json:
        status: "success"
        message: "Operation successful"
        meta:
          pagination:
            has_next: false
            has_prev: false
            limit: 10
            page: 1
            total_items: 10
            total_pages: 1
        data:
          - id: 1
            name: "zambak"
            email: "zamzam@gmail.com"
            phone: "2908763453"
          - id: 3
            name: "foxy"
            email: "foxy@gmail.com"
            phone: "2900643453"
...
  CreateServiceTicketResponse:
    type: object
    properties:
      id:
        type: integer
        example: 8
      vin:
        type: string
        example: "ABC9876543210"
      work_summary:
        type: string
        example: "Brake inspection and pad replacement"
      status:
        type: string
        enum: [open, in_progress, closed]
        example: "open"
      created_at:
        type: string
        format: date-time
        example: "2025-05-21T00:12:03.807165"
      closed_at:
        type: string
        format: date-time
        nullable: true
        example: null
      cost:
        type: number
        format: float
        example: 0.0
      is_deleted:
        type: boolean
        example: false
      customer:
        type: object
        properties:
          id:
            type: integer
            example: 1
          name:
            type: string
            example: "zambak"
          email:
            type: string
            example: "zamzam@gmail.com"
      employees:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
              example: 3
            name:
              type: string
              example: "lilo"
            role:
              type: string
              example: "manager"
      services:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
              example: 3
            service_type:
              type: string
              example: "Oil Change"
            base_price:
              type: string
              example: "19.99"
      serialized_parts:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
              example: 3
            serial_number:
              type: string
              example: "BT-3501"
            status:
              type: string
              enum: [available, used, defective]
              example: "available"
            inventory:
              type: object
              properties:
                id:
                  type: integer
                  example: 5
                name:
                  type: string
                  example: "Spark Plug"
                inventory_number:
                  type: string
                  example: "SP-200"
                price:
                  type: string
                  example: "14.99"
                desc:
                  type: string
                  example: "Spark plug forengines"
...